Fix kern

+++ partitura/score.py
@@ -2058,7 +2058,7 @@ class Measure(TimedObject):
 
     Attributes
     ----------
-    number : int
+    number : intp
         See parameters
 
     """
